// packages/example-data/src/zero/schema.ts
const albumTable = {
  name: "album",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    artistId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "artist_id",
    },
    title: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    year: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["id"],
} as const;
const artistTable = {
  name: "artist",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    sortName: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "sort_name",
    },
    type: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    popularity: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["id"],
} as const;
const cartItemTable = {
  name: "cartItem",
  columns: {
    userId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "user_id",
    },
    albumId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "album_id",
    },
    addedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "added_at",
    },
  },
  primaryKey: ["userId", "albumId"],
  serverName: "cart_item",
} as const;
const albumRelationships = {
  artist: [
    { sourceField: ["artistId"], destField: ["id"], destSchema: "artist", cardinality: "one" },
  ],
  cartItems: [
    { sourceField: ["id"], destField: ["albumId"], destSchema: "cartItem", cardinality: "many" },
  ],
} as const;
const artistRelationships = {
  albums: [
    { sourceField: ["id"], destField: ["artistId"], destSchema: "album", cardinality: "many" },
  ],
} as const;
const cartItemRelationships = {
  album: [{ sourceField: ["albumId"], destField: ["id"], destSchema: "album", cardinality: "one" }],
} as const;
/**
 * The Zero schema object.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const schema = {
  tables: {
    album: albumTable,
    artist: artistTable,
    cartItem: cartItemTable,
  },
  relationships: {
    album: albumRelationships,
    artist: artistRelationships,
    cartItem: cartItemRelationships,
  },
  enableLegacyQueries: false,
  enableLegacyMutators: false,
} as const;
// ...
export const zql = createBuilder(schema);